Implement socketpair for windows
