+++ src/io_engine_select.cpp
@@ -5,6 +5,7 @@ namespace asyncpp::io::detail {
 	std::unique_ptr<io_engine> create_io_engine_select() { return nullptr; }
 } // namespace asyncpp::io::detail
 #else
+#include "io_engine_generic_unix.h"
 
 #include <cstring>
 #include <mutex>
@@ -56,7 +57,7 @@ namespace asyncpp::io::detail {
 		};
 	} // namespace
 
-	class io_engine_select : public io_engine {
+	class io_engine_select : public io_engine_generic_unix {
 	public:
 		io_engine_select();
 		io_engine_select(const io_engine_select&) = delete;
@@ -68,16 +69,8 @@ namespace asyncpp::io::detail {
 		size_t run(bool nowait) override;
 		void wake() override;
 
-		socket_handle_t socket_create(address_type domain, socket_type type) override;
 		void socket_register(socket_handle_t socket) override;
 		void socket_release(socket_handle_t socket) override;
-		void socket_close(socket_handle_t socket) override;
-		void socket_bind(socket_handle_t socket, endpoint ep) override;
-		void socket_listen(socket_handle_t socket, size_t backlog) override;
-		endpoint socket_local_endpoint(socket_handle_t socket) override;
-		endpoint socket_remote_endpoint(socket_handle_t socket) override;
-		void socket_enable_broadcast(socket_handle_t socket, bool enable) override;
-		void socket_shutdown(socket_handle_t socket, bool receive, bool send) override;
 		bool enqueue_connect(socket_handle_t socket, endpoint ep, completion_data* cd) override;
 		bool enqueue_accept(socket_handle_t socket, completion_data* cd) override;
 		bool enqueue_recv(socket_handle_t socket, void* buf, size_t len, completion_data* cd) override;
@@ -87,11 +80,8 @@ namespace asyncpp::io::detail {
 		bool enqueue_send_to(socket_handle_t socket, const void* buf, size_t len, endpoint dst,
 							 completion_data* cd) override;
 
-		file_handle_t file_open(const char* filename, std::ios_base::openmode mode) override;
 		void file_register(file_handle_t fd) override;
 		void file_release(file_handle_t fd) override;
-		void file_close(file_handle_t fd) override;
-		uint64_t file_size(file_handle_t fd) override;
 		bool enqueue_readv(file_handle_t fd, void* buf, size_t len, uint64_t offset, completion_data* cd) override;
 		bool enqueue_writev(file_handle_t fd, const void* buf, size_t len, uint64_t offset,
 							completion_data* cd) override;
@@ -302,101 +292,9 @@ namespace asyncpp::io::detail {
 #endif
 	}
 
-	io_engine::socket_handle_t io_engine_select::socket_create(address_type domain, socket_type type) {
-		int afdomain = -1;
-		switch (domain) {
-		case address_type::ipv4: afdomain = AF_INET; break;
-		case address_type::ipv6: afdomain = AF_INET6; break;
-		case address_type::uds: afdomain = AF_UNIX; break;
-		}
-		int stype = -1;
-		switch (type) {
-		case socket_type::stream: stype = SOCK_STREAM; break;
-		case socket_type::dgram: stype = SOCK_DGRAM; break;
-		case socket_type::seqpacket: stype = SOCK_SEQPACKET; break;
-		}
-		if (afdomain == -1) throw std::system_error(std::make_error_code(std::errc::not_supported));
-		if (stype == -1) throw std::system_error(std::make_error_code(std::errc::not_supported));
-#ifdef __APPLE__
-		auto fd = ::socket(afdomain, stype, 0);
-		if (fd < 0) throw std::system_error(errno, std::system_category(), "select failed");
-		int flags = fcntl(fd, F_GETFL, 0);
-		if (flags < 0 || fcntl(fd, F_SETFL, flags | O_NONBLOCK) < 0 || fcntl(fd, F_SETFD, FD_CLOEXEC) < 0) {
-			close(fd);
-			throw std::system_error(errno, std::system_category(), "fcntl failed");
-		}
-#else
-		auto fd = ::socket(afdomain, stype | SOCK_CLOEXEC | SOCK_NONBLOCK, 0);
-		if (fd < 0) throw std::system_error(errno, std::system_category(), "select failed");
-#endif
-		if (domain == address_type::ipv6) {
-			int opt = 0;
-			if (setsockopt(fd, IPPROTO_IPV6, IPV6_V6ONLY, &opt, sizeof(opt)) < 0) {
-				close(fd);
-				throw std::system_error(errno, std::system_category(), "setsockopt failed");
-			}
-		}
-		return fd;
-	}
-
 	void io_engine_select::socket_register(socket_handle_t socket) {}
 
 	void io_engine_select::socket_release(socket_handle_t socket) {}
-
-	void io_engine_select::socket_close(socket_handle_t socket) {
-		if (socket >= 0) close(socket);
-	}
-
-	void io_engine_select::socket_bind(socket_handle_t socket, endpoint ep) {
-		auto sa = ep.to_sockaddr();
-		auto res = ::bind(socket, reinterpret_cast<sockaddr*>(&sa.first), sa.second);
-		if (res < 0) throw std::system_error(errno, std::system_category(), "bind failed");
-	}
-
-	void io_engine_select::socket_listen(socket_handle_t socket, size_t backlog) {
-		if (backlog == 0) backlog = 20;
-		auto res = ::listen(socket, backlog);
-		if (res < 0) throw std::system_error(errno, std::system_category(), "listen failed");
-	}
-
-	endpoint io_engine_select::socket_local_endpoint(socket_handle_t socket) {
-		sockaddr_storage sa;
-		socklen_t sa_size = sizeof(sa);
-		auto res = getsockname(socket, reinterpret_cast<sockaddr*>(&sa), &sa_size);
-		if (res >= 0) return endpoint(sa, sa_size);
-		throw std::system_error(errno, std::system_category(), "getsockname failed");
-	}
-
-	endpoint io_engine_select::socket_remote_endpoint(socket_handle_t socket) {
-		sockaddr_storage sa;
-		socklen_t sa_size = sizeof(sa);
-		auto res = getpeername(socket, reinterpret_cast<sockaddr*>(&sa), &sa_size);
-		if (res >= 0)
-			return endpoint(sa, sa_size);
-		else if (res < 0 && errno != ENOTCONN)
-			throw std::system_error(errno, std::system_category(), "getpeername failed");
-		return {};
-	}
-
-	void io_engine_select::socket_enable_broadcast(socket_handle_t socket, bool enable) {
-		int opt = enable ? 1 : 0;
-		auto res = setsockopt(socket, SOL_SOCKET, SO_BROADCAST, &opt, sizeof(opt));
-		if (res < 0) throw std::system_error(errno, std::system_category(), "setsockopt failed");
-	}
-
-	void io_engine_select::socket_shutdown(socket_handle_t socket, bool receive, bool send) {
-		int mode = 0;
-		if (receive && send)
-			mode = SHUT_RDWR;
-		else if (receive)
-			mode = SHUT_RD;
-		else if (send)
-			mode = SHUT_WR;
-		else
-			return;
-		auto res = ::shutdown(socket, mode);
-		if (res < 0 && errno != ENOTCONN) throw std::system_error(errno, std::system_category(), "shutdown failed");
-	}
 
 	bool io_engine_select::enqueue_connect(socket_handle_t socket, endpoint ep, completion_data* cd) {
 		auto sa = ep.to_sockaddr();
@@ -546,41 +444,9 @@ namespace asyncpp::io::detail {
 		return false;
 	}
 
-	io_engine::file_handle_t io_engine_select::file_open(const char* filename, std::ios_base::openmode mode) {
-		if ((mode & std::ios_base::ate) == std::ios_base::ate) throw std::logic_error("unsupported flag");
-		int m = 0;
-		if ((mode & std::ios_base::app) == std::ios_base::app) m |= O_APPEND;
-		if ((mode & std::ios_base::in) == std::ios_base::in)
-			m |= ((mode & std::ios_base::out) == std::ios_base::out) ? O_RDWR : O_RDONLY;
-		else if ((mode & std::ios_base::out) == std::ios_base::out)
-			m |= O_WRONLY;
-		else
-			throw std::invalid_argument("neither std::ios::in, nor std::ios::out was specified");
-		if ((mode & std::ios_base::trunc) == std::ios_base::trunc) m |= O_TRUNC;
-		auto res = ::open(filename, m, 0660);
-		if (res < 0) throw std::system_error(errno, std::system_category());
-		return res;
-	}
-
 	void io_engine_select::file_register(file_handle_t fd) {}
 
 	void io_engine_select::file_release(file_handle_t fd) {}
-
-	void io_engine_select::file_close(file_handle_t fd) { ::close(fd); }
-
-	uint64_t io_engine_select::file_size(file_handle_t fd) {
-#ifdef __APPLE__
-		struct stat info {};
-		auto res = fstat(fd, &info);
-		if (res < 0) throw std::system_error(errno, std::system_category());
-		return info.st_size;
-#else
-		struct stat64 info {};
-		auto res = fstat64(fd, &info);
-		if (res < 0) throw std::system_error(errno, std::system_category());
-		return info.st_size;
-#endif
-	}
 
 	bool io_engine_select::enqueue_readv(file_handle_t fd, void* buf, size_t len, uint64_t offset,
 										 completion_data* cd) {
